API hooked up upto Publications. Need to test it though. Also need to reset input field errors for all forms.

+++ app/Http/Controllers/PublicationController.php
@@ -46,7 +46,9 @@ class PublicationController extends Controller
 
     public function show($id)
     {
-        $publication = Publication::where('id', $id)->select('id', 'publication_type_info_id', 'publication_name', 'publication_file', 'created_at')->first();
+        $publication = Publication::where('id', $id)->select('id', 'publication_type_info_id', 'publication_name', 'publication_file', 'created_at')->with(['publication_type_info' => function($query){
+            return $query->select(['id', 'publication_type_name']);
+        }])->first();
         if($publication){
             return response()->json($publication, 200);
         }
@@ -59,32 +61,41 @@ class PublicationController extends Controller
         $request->validate([
             'publication_type_info_id' => ['required','numeric', 'integer', 'exists:publication_type_infos,id'],
             'publication_name' => ['required', 'string', 'max:255'],
-            'publication_file' => ['required', 'file', 'mimetypes:application/pdf', 'max:50000']
         ]);
 
         $publication = Publication::find($id);
-
-        //get filename with extension
-        $filenameWithExt = $request->file('publication_file')->getClientOriginalName();
-        //get just file name (using standard php function)
-        $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
-        //get just extension
-        $extension = $request->file('publication_file')->getClientOriginalExtension();
-        //filename to store(uses a time php function to get current time)
-        //this string is a unique name so that file with duplicate name do not get uploaded and
-        //cause problems when viewing(same problem that occured in CISV photo gallery)
-        $filenameToStore= $filename.'_'.time().'.'.Str::lower($extension);
-        //upload image
-        $request->file('publication_file')->storeAs('public/publication_files', $filenameToStore);
-
         if($publication){
+
+            if($request->hasFile('publication_file')) {
+
+                $request->validate([
+                    'publication_file' => ['required', 'file', 'mimetypes:application/pdf', 'max:50000']
+                ]);
+                
+                Storage::delete('public/publication_files/'.$publication->publication_file);
+                //get filename with extension
+                $filenameWithExt = $request->file('publication_file')->getClientOriginalName();
+                //get just file name (using standard php function)
+                $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+                //get just extension
+                $extension = $request->file('publication_file')->getClientOriginalExtension();
+                //filename to store(uses a time php function to get current time)
+                //this string is a unique name so that file with duplicate name do not get uploaded and
+                //cause problems when viewing(same problem that occured in CISV photo gallery)
+                $filenameToStore= $filename.'_'.time().'.'.Str::lower($extension);
+                //upload image
+                $request->file('publication_file')->storeAs('public/publication_files', $filenameToStore);
+            }
+            else{
+                $filenameToStore = $publication->publication_file;
+            }
+
             $publication->publication_type_info_id = $request->publication_type_info_id;
-            $publication->publication_name = $request->publication_name;            
-            Storage::delete('public/publication_files/'.$publication->publication_file);
+            $publication->publication_name = $request->publication_name;
             $publication->publication_file = $filenameToStore;
             $publication->save();
     
-            return response()->json('Publication Updated Successfully', 201);
+            return response()->json('Publication Updated Successfully !!', 201);
         }
 
         return response()->json('Publication With ID Not Found !!', 404);
@@ -98,7 +109,7 @@ class PublicationController extends Controller
             Storage::delete('public/publication_files/'.$publication->publication_file);                            
             $publication->delete();
     
-            return response()->json('Publication Deleted Successfully', 201);
+            return response()->json('Publication Deleted Successfully !!', 201);
         }
 
         return response()->json('Publication With ID Not Found !!', 404);
